Move 'switch review' page to correct location and replace placeholder
